requirements.txt
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
bcrypt==4.0.1 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
cryptography==39.0.2 ; python_version >= "3.10" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.95.0 ; python_version >= "3.10" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.10" and platform_machine == "aarch64" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "ppc64le" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "x86_64" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "amd64" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "AMD64" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "win32" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "WIN32" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.24.2 ; python_version >= "3.10" and python_version < "4.0"
pandas==1.5.3 ; python_version >= "3.10" and python_version < "4.0"
passlib[bcrypt]==1.7.4 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.6 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-jose[cryptography]==3.3.0 ; python_version >= "3.10" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.10" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy2-stubs==0.0.2a32 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==1.4.41 ; python_version >= "3.10" and python_version < "4.0"
sqlmodel==0.0.8 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.26.1 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.21.1 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.1 ; python_version >= "3.10" and python_version < "4.0"
psycopg2==2.9.1 ; python_version >= "3.10" and python_version < "4.0"